@@ -4,7 +4,7 @@ import android.annotation.TargetApi;
 import android.app.usage.UsageEvents;
 import android.content.pm.PackageManager;
 import android.os.Build;
-import android.support.v7.widget.RecyclerView;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
